requirements.txt
fastapi
uvicorn
pydantic
google-genai
aiohttp
openai
python-dotenv
